記事初期ロードJS削減
- MDXComponents が静的importしていた重量コンポーネントを遅延分離
  - DonutChart/LineChart は next/dynamic(ssr:false) ラッパー(LazyChart)経由にし
    chart.js を初期バンドルから除去
  - Mermaid は本体をマウント時の動的importに変更し mermaid.render API化。
    テーマ切替で図が再描画されないバグも修正
- kbar検索一式を SearchModal に隔離し、検索ボタン/⌘K の初回発火時に遅延ロード。
  /search.json のフェッチも検索を開くまで遅延(従来は全ページで取得)
- 未使用の pliny/search/algolia.css を削除
実測(本番ビルド): 記事ページ初期JS 1,629KB(gzip 473KB) → 748KB(gzip 241KB)
トップページ 757KB(gzip 239KB) → 694KB(gzip 220KB)
