+++ app/layout.tsx
@@ -1,5 +1,4 @@
 import 'css/tailwind.css'
-import 'pliny/search/algolia.css'
 import 'remark-github-blockquote-alert/alert.css'
 
 import { ReactNode } from 'react'
