+++ components/MDXComponents.tsx
@@ -12,8 +12,7 @@ import SpotifyPlayer from '@/components/Spotify'
 import AppleMusicPlayer from '@/components/AppleMusic'
 import { Tweet } from 'react-tweet'
 import TableWrapper from './TableWrapper'
-import DonutChart from './DonutChart'
-import LineChart from './LineChart'
+import { DonutChart, LineChart } from './LazyChart'
 import Mermaid from './Mermaid'
 import BlueskyPostEmbed from '@/components/Bluesky'
 import Dots from '@/components/Dots'
